% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{massage_cuts}
\alias{massage_cuts}
\title{Massage Cuts to the Desired Shape}
\usage{
massage_cuts(cuts, k)
}
\arguments{
\item{cuts}{A matrix, list, or vector of cuts}

\item{k}{Optional \code{k} saying how many times the vector of cuts should be
repeated. Only matters when \code{cuts} is a vector.}
}
\description{
Massage Cuts to the Desired Shape
}
