% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ordered_y}
\alias{ordered_y}
\title{Transform y Into a Form Where Each Category is An Positive Integer}
\usage{
ordered_y(y)
}
\arguments{
\item{y}{An array or data frame of observations.}
}
\value{
A matrix.
}
\description{
Transform y Into a Form Where Each Category is An Positive Integer
}
\keyword{internal}
