% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trim_vector}
\alias{trim_vector}
\title{Remove Infinities from Vector, Append and Prepend \code{-Inf} and \code{Inf}, and Sort}
\usage{
trim_vector(x)
}
\arguments{
\item{x}{Numeric vector with no \code{NA} values.}
}
\value{
Vector with \verb{-Inf`` prepended and }Inf` appended
}
\description{
Remove Infinities from Vector, Append and Prepend \code{-Inf} and \code{Inf}, and Sort
}
\keyword{internal}
