% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{x_hat}
\alias{x_hat}
\title{Transform Likert data to X_hats.}
\usage{
x_hat(y, cuts)
}
\arguments{
\item{y}{Vector of observations.}

\item{cuts}{Vector of cuts.}
}
\value{
The X_hats associated with \code{y} and \code{cuts}.
}
\description{
Transform Likert data to X_hats.
}
\keyword{internal}
