% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xi_sample}
\alias{xi_sample}
\title{Calculate Sample Xi}
\usage{
xi_sample(y, cuts, use = "complete.obs")
}
\arguments{
\item{y}{An array or data frame of observations.}

\item{cuts}{A matrix, list, or vector of cuts.}

\item{use}{Passed to \code{stats::cov}; defaults to \code{"complete.obs"}.}
}
\value{
Sample xi matrix.
}
\description{
Calculate Sample Xi
}
\keyword{internal}
