% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xi_theoretical}
\alias{xi_theoretical}
\title{Calculate the Theoretical Xi}
\usage{
xi_theoretical(cuts, rho)
}
\arguments{
\item{cuts}{A matrix, list, or vector of cuts.}

\item{rho}{The polychoric correlation matrix; a numeric matrix with no
missing entries.}
}
\value{
Theoretical xi matrix.
}
\description{
Calculate the Theoretical Xi
}
\keyword{internal}
